refactor(PowerDisplay): hoist MonitorDisplayInfo to shared namespace, add OutputTechnology/IsInternal fields
